Module Srk.ExpPolynomial

type t

val pp : Stdlib.Format.formatter -> t -> unit

val show : t -> string

val equal : t -> t -> bool

val add : t -> t -> t

val mul : t -> t -> t

val negate : t -> t

val zero : t

val one : t

val of_polynomial : Polynomial.QQX.t -> t

val of_exponential : QQ.t -> t

val scalar : QQ.t -> t

val scalar_mul : QQ.t -> t -> t

val compose_left_affine : t -> int -> int -> t

compose_left_affine f a b computes the function lambda x. f (ax + b)

val summation : t -> t

summation f computes an exponential-polynomial g such that g(n) = sum_{i=0}^n f(i).

val solve_rec : ?⁠initial:QQ.t -> QQ.t -> t -> t

solve_rec ?initial lambda g computes an exponential-polynomial g such that

g(0) = initial

g(n+1) = lambda*g(n) + f(n)

. If the initial value is omitted, it is taken to be 0.

val term_of : 'a Syntax.context -> 'a Syntax.term -> t -> 'a Syntax.term

term_of srk t f computes a term representing f(t).

val eval : t -> int -> QQ.t

val enum : t -> (Polynomial.QQX.t * QQ.t) BatEnum.t

val add_term : Polynomial.QQX.t -> QQ.t -> t -> t

val of_term : Polynomial.QQX.t -> QQ.t -> t

module Vector : sig ... end

Vectors with exponential-polynomial entries

module Matrix : sig ... end

Matrices with exponential-polynomial entries

val exponentiate_rational : Linear.QQMatrix.t -> Matrix.t option

Symbolically exponentiate a matrix with rational eigenvalues. None indicates the matrix has irrational eigenvalues.

module UltPeriodic : sig ... end