Module ExpPolynomial.UltPeriodic

type elt = t

type t

val equal : t -> t -> bool

val add : t -> t -> t

val negate : t -> t

val zero : t

val mul : t -> t -> t

val one : t

val pp : Stdlib.Format.formatter -> t -> unit

val show : t -> string

val make : QQ.t list -> elt list -> t

make t p constructs an ultimately periodic sequence

val transient : t -> QQ.t list

Retrieve the transient part of a sequence

val periodic : t -> elt list

Retrieve the periodic part of a sequence

val enum : t -> QQ.t BatEnum.t

Enumerate the sequence.

val eval : t -> int -> QQ.t

val period_len : t -> int

val transient_len : t -> int

val summation : t -> t

Given a function f, compute the function lambda n. sum_{i=0}^n f(i)

val solve_rec : ?⁠initial:QQ.t -> QQ.t -> t -> t

solve_rec ?initial coeff f computes the function g that satisfies

g(0) = initial

g(n+1) = coeff*g(n) + f_n(n)

.

If initial is omitted, it is taken to be 0.

val term_of : 'a Syntax.context -> 'a Syntax.term -> 'a Syntax.term -> t -> 'a Syntax.term

term_of srk q r f computes a term representing f(qp + r), where p is the period of f.

val scalar : QQ.t -> t

val of_polynomial : Polynomial.QQX.t -> t

val of_exponential : QQ.t -> t

val of_exp_polynomial : elt -> t

val compose_left_affine : t -> int -> int -> t

compose_left_affine f a b computes an ultimately periodic function g such that for all k g_i(i) = f_{ai+b}(ai+b)

val flatten : t list -> t

flatten f_0 ... f_{p-1} computes a function g such that g(qp + r) = f_r(q)

val shift : QQ.t list -> t -> t

flatten [t_0, ..., t_{n-1}] f computes a function g such that g(i) = t_i for i < n and g(i) = f(i-n) otherwise. If we imagine f as an infinite sequenc f_0 f_1 f_2 ..., then g is the sequence t_0, ..., t_{n-1}, f_0, f_1, f_2, ....