Module Srk.Iteration

module type PreDomain = sig ... end

module type PreDomainIter = sig ... end

module type PreDomainWedge = sig ... end

module type PreDomainWedgeIter = sig ... end

module type Domain = sig ... end

module WedgeGuard : PreDomainWedgeIter

module PolyhedronGuard : sig ... end

module LinearGuard : PreDomainIter

module LinearRecurrenceInequation : PreDomain

Abstract a transition formula F(x,x') by a system of recurrences of the form a(x') >= a(x) + c where a is a linear map and c is a scalar.

module NonlinearRecurrenceInequation : PreDomainWedge

Abstract a transition formula F(x,x',y) (where y denotes a set of symbolic constants) by a system of recurrences of the form ax' >= ax + t(y) where a is a linear map and t(y) is a (possibly non-linear) term over y.

module Split : functor (Iter : PreDomain) -> PreDomain

Improve iteration operator using split invariants

module InvariantDirection : functor (Iter : PreDomain) -> PreDomain

Improve iteration operator using variable directions (increasing, decreasing, stable) to find phase splits.

module Product : functor (A : PreDomain) -> functor (B : PreDomain) -> PreDomain with type 'a t = 'a A.t * 'a B.t

module ProductWedge : functor (A : PreDomainWedge) -> functor (B : PreDomainWedge) -> PreDomainWedge with type 'a t = 'a A.t * 'a B.t

Same as product, but faster for iteration domains that abstract through the wedge domain.

module MakeDomain : functor (Iter : PreDomain) -> Domain