Module Iteration.LinearRecurrenceInequation

Abstract a transition formula F(x,x') by a system of recurrences of the form a(x') >= a(x) + c where a is a linear map and c is a scalar.

type 'a t

val pp : 'a Syntax.context -> (Syntax.symbol * Syntax.symbol) list -> Stdlib.Format.formatter -> 'a t -> unit

val exp : 'a Syntax.context -> (Syntax.symbol * Syntax.symbol) list -> 'a Syntax.term -> 'a t -> 'a Syntax.formula

val abstract : 'a Syntax.context -> 'a TransitionFormula.t -> 'a t