Module Lts.PartialLinearMap

Partial linear maps

type t

A partial linear map is a partial function that is linear and whose domain is a vector space

val equal : t -> t -> bool

val identity : int -> t

identity n is the (total) identity map on the coordinates 0 .. n-1; all other coordinates are sent to 0.

val make : QQMatrix.t -> QQVector.t list -> t

make A G creates the partial map whose action is given by the matrix A, and whose domain of definition is the set of all v such that g.v = 0 for all g in G. (I.e., G is a set of equality constraints that must on the input for make A G to be defined; the domain is the orthogonal complement of G).

val pp : Stdlib.Format.formatter -> t -> unit

val compose : t -> t -> t

(Partial) map composition

val iteration_sequence : t -> t list * QQVector.t list

Repeatly compose a partial map with itself until the domain stabilizes; return both iteration sequence and the invariant guard of the function.

We have guard(f) <= guard(f o f) <= guard(f o f o f) <= ... and there is some (least) n such that guard(f^n) = guard(f^m) for all m >= n. The pair returned by iteration_sequence f is ([f; f o f; ...; f^n], dom(f^n)).

val map : t -> QQMatrix.t

Access the underlying (total) map of a partial map.

val guard : t -> QQVectorSpace.t

Access the guard of a partial map. The domain of the map is its orthogonal complement