Module Srk.Nonlinear

val ensure_symbols : 'a Syntax.context -> unit

Ensure that a context has the following named symbols

mul (real multiplication)

imul (integer multiplication)

mod (real modulo)

imod (integer modulo)

inv (real inverse).

pow (real exponentiation)

log (real logarithm) If the symbols are not present in the context ensure_symbols registers them.

val uninterpret_rewriter : 'a Syntax.context -> ('a, Syntax.typ_fo) Syntax.expr -> ('a, Syntax.typ_fo) Syntax.expr

Replace non-linear arithmetic with uninterpreted functions. The uninterpreted function symbols are named symbols: mul, div, and mod. This rewriter is safe to apply top-down or bottom-up.

val interpret_rewriter : 'a Syntax.context -> ('a, Syntax.typ_fo) Syntax.expr -> ('a, Syntax.typ_fo) Syntax.expr

Replace non-linear uninterpreted functions with interpreted ones. This rewriter is safe to apply top-down or bottom-up.

val uninterpret : 'a Syntax.context -> ('a, 'b) Syntax.expr -> ('a, 'b) Syntax.expr

Replace non-linear arithmetic with uninterpreted functions. The uninterpreted function symbols are named symbols: mul, div, and mod.

val interpret : 'a Syntax.context -> ('a, 'b) Syntax.expr -> ('a, 'b) Syntax.expr

Replace non-linear uninterpreted functions with interpreted ones.

val linearize : 'a Syntax.context -> 'a Syntax.formula -> 'a Syntax.formula

Compute a linear approximation of a non-linear formula.

val mk_log : 'a Syntax.context -> 'a Syntax.term -> 'a Syntax.term -> 'a Syntax.term

val mk_pow : 'a Syntax.context -> 'a Syntax.term -> 'a Syntax.term -> 'a Syntax.term

val optimize_box : ?⁠context:SrkZ3.z3_context -> 'a Syntax.context -> 'a Syntax.formula -> 'a Syntax.term list -> [ `Sat of Interval.t list | `Unsat | `Unknown ]

Given a formula phi and a list of objectives o1,...,on, find the least bounding interval for each objective within the feasible region defined by the formula.

val simplify_terms_rewriter : 'a Syntax.context -> 'a Syntax.rewriter

Simplification rules power and log.

val simplify_terms : 'a Syntax.context -> 'a Syntax.formula -> 'a Syntax.formula

Simplify power and log terms.

val simplify_term : 'a Syntax.context -> 'a Syntax.term -> 'a Syntax.term

Simplify power and log terms.