Module Polynomial.Ideal

A polynomial ideal is a set of polynomials that is closed under addition, and is closed under multiplication by any polynomial.

type t

val pp : (Stdlib.Format.formatter -> int -> unit) -> Stdlib.Format.formatter -> t -> unit

Pretty print

val make : QQXs.t list -> t

Compute the smallest ideal that contains a given set of polynomials

val generators : t -> QQXs.t list

Compute a finite set of polynomials that generates the given ideal. Note make (generators i) = i, but generators (make g) is not necessarily equal to g.

val subset : t -> t -> bool

Is one ideal contained inside another?

val equal : t -> t -> bool

Is one ideal equal to another?

val mem : QQXs.t -> t -> bool

Does an ideal contain a given polynomial?

val intersect : t -> t -> t

Intersect two ideals.

val product : t -> t -> t

Compute the ideal consisting of all products of polynomials belonging to the two given ideals

val sum : t -> t -> t

Compute the ideal consisting of all sums of polynomials beloning to the two given ideals

val project : (int -> bool) -> t -> t

Compute the ideal consisting of all polynomials in the given ideal that are defined only over dimensions satisfying the given predicate.