Module Ring.MakeUltPeriodicSeq

Ultimately periodic sequences. An ultimately periodic sequence is an infinite sequence consisting of a finite transient phase and an infinite periodic phase (with finite period). Ring operations are pointwise:

(f+g)(i) = f(i) + g(i)

(f*g)(i) = f(i) * g(i) (Hadamard product)

(-f)(i) = -f(i)

(0)(i) = 0

(1)(i) = 1

Parameters

R : S

Signature

include S

type t

val equal : t -> t -> bool

val add : t -> t -> t

val negate : t -> t

val zero : t

val mul : t -> t -> t

val one : t

val pp : (Stdlib.Format.formatter -> R.t -> unit) -> Stdlib.Format.formatter -> t -> unit

val make : R.t list -> R.t list -> t

make t p constructs an ultimately periodic sequence

val transient : t -> R.t list

Retrieve the transient part of a sequence

val periodic : t -> R.t list

Retrieve the periodic part of a sequence

val enum : t -> R.t BatEnum.t

Enumerate the sequence.

val period_len : t -> int

Length of the period of a sequence

val transient_len : t -> int

Length of the transient part of a sequence