Module Srk.SrkZ3

type z3_context = Z3.context

type z3_expr = Z3.Expr.expr

type z3_func_decl = Z3.FuncDecl.func_decl

type 'a open_expr = [

| `Real of QQ.t

| `App of z3_func_decl * 'a list

| `Var of int * Syntax.typ_fo

| `Add of 'a list

| `Mul of 'a list

| `Binop of [ `Div | `Mod ] * 'a * 'a

| `Unop of [ `Floor | `Neg ] * 'a

| `Tru

| `Fls

| `And of 'a list

| `Or of 'a list

| `Not of 'a

| `Ite of 'a * 'a * 'a

| `Quantify of [ `Exists | `Forall ] * string * Syntax.typ_fo * 'a

| `Atom of [ `Eq | `Leq | `Lt ] * 'a * 'a

]

val z3_of_term : 'a Syntax.context -> z3_context -> 'a Syntax.term -> z3_expr

val z3_of_formula : 'a Syntax.context -> z3_context -> 'a Syntax.formula -> z3_expr

val z3_of_expr : 'a Syntax.context -> z3_context -> ('a, Syntax.typ_fo) Syntax.expr -> z3_expr

val term_of_z3 : 'a Syntax.context -> z3_expr -> 'a Syntax.term

Convert a Z3 expression into a term. Raises Invalid_argument on failure.

val formula_of_z3 : 'a Syntax.context -> z3_expr -> 'a Syntax.formula

Convert a Z3 expression into a formula. Raises Invalid_argument on failure.

val expr_of_z3 : 'a Syntax.context -> z3_expr -> ('a, Syntax.typ_fo) Syntax.expr

Convert a Z3 expression into an expression. Raises Invalid_argument on failure.

module Solver : sig ... end

val mk_solver : ?⁠context:z3_context -> ?⁠theory:string -> 'a Syntax.context -> 'a Solver.t

val optimize_box : ?⁠context:z3_context -> 'a Syntax.context -> 'a Syntax.formula -> 'a Syntax.term list -> [ `Sat of Interval.t list | `Unsat | `Unknown ]

Given a formula phi and a list of objectives o1,...,on, find the least bounding interval for each objective within the feasible region defined by the formula.

val qe : ?⁠context:z3_context -> 'a Syntax.context -> 'a Syntax.formula -> 'a Syntax.formula

Quantifier elimination

val simplify : ?⁠context:z3_context -> 'a Syntax.context -> 'a Syntax.formula -> 'a Syntax.formula

Simplify a formula using Z3's routines. The resulting formula is equisatisfiable, but not necessarily equivalent, to the original.

val load_smtlib2 : ?⁠context:z3_context -> 'a Syntax.context -> string -> 'a Syntax.formula

Parse a SMTLIB2-formatted string

val interpolate_seq : ?⁠context:z3_context -> 'a Syntax.context -> 'a Syntax.formula list -> [ `Sat of 'a Interpretation.interpretation | `Unsat of 'a Syntax.formula list | `Unknown ]

Sequence interpolation. Returns either a model of the formula or a sequence of interpolants if there is no model.

module CHC : sig ... end

Constrained Horn Clauses