Module Syntax.ContextTable

A context table is a hash table mapping contents to values. If a context is garbage collected, the corresponding entry in the table will be removed. The values stored in a context table should have pointers back to their associated context.

type 'a t

val create : int -> 'a t

val add : 'a t -> 'b context -> 'a -> unit

val replace : 'a t -> 'b context -> 'a -> unit

val find : 'a t -> 'b context -> 'a

val mem : 'a t -> 'b context -> bool

val clear : 'a t -> unit