Module Srk.TransitionFormula

type 'a t

A transition formula is a logical formula that describes a binary relation on a set of states. Transition formulas are assumed to be quantifier-free, and are over four sets of symbols: pre-state symbols, post-state symbols symbolic constants, and Skolem constants. Pre-state and post-state symbols are in 1-1 correspondence and define the space on which the transition formula acts. Symbolic constants are parameters that do not change during a computation; wheras Skolem constants do (and should be thought of as existentially quantified variables).

val make : ?⁠exists:(Syntax.symbol -> bool) -> 'a Syntax.formula -> (Syntax.symbol * Syntax.symbol) list -> 'a t

val identity : 'a Syntax.context -> (Syntax.symbol * Syntax.symbol) list -> 'a t

Identity relation

val mul : 'a Syntax.context -> 'a t -> 'a t -> 'a t

Sequential composition

val add : 'a Syntax.context -> 'a t -> 'a t -> 'a t

Union

val formula : 'a t -> 'a Syntax.formula

Retrieve the underlying formula

val symbols : 'a t -> (Syntax.symbol * Syntax.symbol) list

Retrieve the pre-state/post-state symbols of a transition formula

val exists : 'a t -> Syntax.symbol -> bool

Retrieve the predicate that identifies skolem constants

val is_symbolic_constant : 'a t -> Syntax.symbol -> bool

Retrieve a predicate that identifies symbolic constants

val symbolic_constants : 'a t -> Srk.Syntax.Symbol.Set.t

Retrieve the set of symbolic constants that appear in the underyling formula

val wedge_hull : 'a Syntax.context -> 'a t -> 'a Wedge.t

Find a wedge over the pre-state, post-state, and constant symbols that over-approximates the underlying formula of a transition formula

val linearize : 'a Syntax.context -> 'a t -> 'a t

Overapproxmate a transition formula by a linear arithmetic transition formula

val pre_map : 'a Syntax.context -> (Syntax.symbol * Syntax.symbol) list -> 'a Syntax.term Srk.Syntax.Symbol.Map.t

Map pre-state symbols to their post-state counterparts

val post_map : 'a Syntax.context -> (Syntax.symbol * Syntax.symbol) list -> 'a Syntax.term Srk.Syntax.Symbol.Map.t

Map post-state symbols to their pre-state counterparts

val pre_symbols : (Syntax.symbol * Syntax.symbol) list -> Srk.Syntax.Symbol.Set.t

Retrieve the set of pre-state symbols

val post_symbols : (Syntax.symbol * Syntax.symbol) list -> Srk.Syntax.Symbol.Set.t

Retrieve the set of post-state symbols