Module Srk.WeightedGraph

type 'a weighted_graph

type 'a t = 'a weighted_graph

type 'a algebra = {

mul : 'a -> 'a -> 'a;

add : 'a -> 'a -> 'a;

star : 'a -> 'a;

zero : 'a;

one : 'a;

}

Regular algebra signature

type ('a, 'b) omega_algebra = {

omega : 'a -> 'b;

omega_add : 'b -> 'b -> 'b;

omega_mul : 'a -> 'b -> 'b;

}

Omega algebra signature

module U : Graph.Sig.G with type U.V.t = int

Unweighted graphs

type vertex = int

val empty : 'a algebra -> 'a t

Create an empty weighted graph over the given algebra of weights.

val add_vertex : 'a t -> vertex -> 'a t

Add a vertex to a graph.

val add_edge : 'a t -> vertex -> 'a -> vertex -> 'a t

add_edge g u w v adds an edge u -w-> v to g. If there is already an edge from u to v, then w is added to the weight of the edge.

val edge_weight : 'a t -> vertex -> vertex -> 'a

Find the weight associated with a single edge.

val path_weight : 'a t -> vertex -> vertex -> 'a

path_weight g u v computes the sum of all weighted paths from u to v in g.

val msat_path_weight : 'a t -> vertex list -> 'a t

Multiple-source all-target path expressions. msat_path_weight g srcs computes a function f which maps a vertex u belonging to src and any other vertex v to to the sum of all weighted paths from u to v.

val omega_path_weight : 'a t -> ('a, 'b) omega_algebra -> vertex -> 'b

omega_path_weight g alg v computes the sum of all weighted omega paths starting at v in g.

val cut_graph : 'a t -> vertex list -> 'a t

cut_graph g c computes the cut graph g' g' = <c, { (u, w, v) : (u, v) in c x c, w is the sum of all paths from u to v in g not going through any node in c }> *

val remove_vertex : 'a t -> vertex -> 'a t

Remove a vertex from a graph.

val contract_vertex : 'a t -> vertex -> 'a t

contract g v removes vertex v from g while preserving all weighted paths among remaining vertices. That is, for each pair of edges p -pw-> v and v -sw-> s, adds the edge p -> s with weight pw.vw*.sw, where vw is the weight of the self-loop at v.

val split_vertex : 'a t -> vertex -> 'a -> vertex -> 'a t

val forget_weights : 'a t -> U.t

val map_weights : (vertex -> 'a -> vertex -> 'a) -> 'a t -> 'a t

val fold_edges : ((vertex * 'a * vertex) -> 'b -> 'b) -> 'a t -> 'b -> 'b

val iter_edges : ((vertex * 'a * vertex) -> unit) -> 'a t -> unit

val fold_succ_e : ((vertex * 'a * vertex) -> 'b -> 'b) -> 'a t -> vertex -> 'b -> 'b

val fold_pred_e : ((vertex * 'a * vertex) -> 'b -> 'b) -> 'a t -> vertex -> 'b -> 'b

val iter_succ_e : ((vertex * 'a * vertex) -> unit) -> 'a t -> vertex -> unit

val iter_pred_e : ((vertex * 'a * vertex) -> unit) -> 'a t -> vertex -> unit

val fold_vertex : (vertex -> 'b -> 'b) -> 'a t -> 'b -> 'b

val iter_vertex : (vertex -> unit) -> 'a t -> unit

val mem_edge : 'a t -> vertex -> vertex -> bool

val max_vertex : 'a t -> int

Find the largest integer identifier for a vertex in a graph

val forward_analysis : 'a t -> entry:int -> update:(pre:'v -> 'a -> post:'v -> 'v option) -> init:(int -> 'v) -> int -> 'v

Compute least fixpoint solution to a system of constraints defined over a weighted graph. The type 'v represents values in poset. We compute the least annotation inv of the graph such that

inv(v) >= init(v) for all vertices v

update (inv src) weight (inv dst) = None for each edge (src, weight, dst)

The update function associates each edge with a constraint. The constraint is satisfied when the update function returns None; otherwise update pre weight post should return the least value post' such that

post' >= post

update pre weight post' = None (i.e., post' satisfies the associated constraint)

The poset 'v is expected to satisfy the ascending chain condition, and update is expected to be monotone in the sense that if pre >= pre' and update pre weight post = None, then update pre' weight post = None.

module type AbstractWeight = sig ... end

module RecGraph : sig ... end

A recursive graph is a graph with two types of edges: simple edges and call edges. Each call edge designates an entry vertex and exit vertex, and can be interpreted a set of paths that begin and entry and end at exit.