Module Srk__Abstract

val affine_hull : 'a Srk.Syntax.context -> 'a Srk.Syntax.formula -> Srk.Syntax.symbol list -> 'a Srk.Syntax.term list

affine_hull srk phi symbols computes a basis for the affine hull of phi, projected onto the given set of symbols. The basis is represented as a list of terms, with the interpretation that a point p belongs to the affine hull if every term in the list evaluates to 0 at p.

val vanishing_space : 'a Srk.Syntax.context -> 'a Srk.Syntax.formula -> 'a Srk.Syntax.term array -> Srk.Linear.QQVectorSpace.t

Let cs = t1, ..., tn be a list of terms. t1, ..., tn generates a linear space of functions from interpretations to rationals; vanishing_space srk phi cs computes a basis for the subspace of functions from interpretations to rationals that evaluate to zero on all models of phi.

val boxify : 'a Srk.Syntax.context -> 'a Srk.Syntax.formula -> 'a Srk.Syntax.term list -> 'a Srk.Syntax.formula

boxify srk phi terms computes the strongest formula of the form /\ { lo <= t <= hi : t in terms } that is implied by phi.

val abstract : ?⁠exists:(Srk.Syntax.symbol -> bool) -> 'a Srk.Syntax.context -> 'abs Apron.Manager.t -> 'a Srk.Syntax.formula -> ('a, 'abs) Srk.SrkApron.property

abstract ?exists srk man phi computes the strongest property that is implied by phi which is expressible within a given abstract domain. The property is restricted to use only the symbols that satisfy the ?exists predicate (which defaults to the constant true predicate).

module Sign : sig ... end

module MakeAbstractRSY : functor (C : sig ... end) -> sig ... end

Symbolic abstraction as described in Reps, Sagiv, Yorsh---"Symbolic implementation of the best transformer", VMCAI 2004.