Module Srk__Abstract.MakeAbstractRSY

Symbolic abstraction as described in Reps, Sagiv, Yorsh---"Symbolic implementation of the best transformer", VMCAI 2004.

Parameters

C : sig ... end

Signature

module type Domain = sig ... end

Domains must satisfy the ascending chain condition, and are equipped with a function of_model that computes the best abstraction of a single model.

module Sign : Domain

Sign analysis determines whether each variables is positive, nonnegative, zero, nonpositive, negative, or unknown.

module AffineRelation : Domain with type t = (C.t, Polka.equalities Polka.t) Srk.SrkApron.property

Domain of affine equalities

module PredicateAbs : functor (U : sig ... end) -> Domain

Predicate abstraction

module Product : functor (A : Domain) -> functor (B : Domain) -> Domain with type t = A.t * B.t

Reduced product of abstract domains

val abstract : ?⁠exists:(Srk.Syntax.symbol -> bool) -> (module Domain with type t = 'a) -> C.t Srk.Syntax.formula -> 'a

Compute the best abstraction of a formula within a given domain. For a formula phi, this is an element of the abstract domain elt such that: (1) phi |= formula_of elt, (2) elt is expressed only over the symbols that satisy the exists predicate, and (3) for any elt' satisfying (1) and (2), we have elt <= elt'.