Module MakeAbstractRSY.AffineRelation

Domain of affine equalities

type t = (C.t, Polka.equalities Polka.t) Srk.SrkApron.property

val top : t

val bottom : t

val exists : (Srk.Syntax.symbol -> bool) -> t -> t

Project property onto the symbols that satisfy the given predicate

val join : t -> t -> t

val equal : t -> t -> bool

val of_model : C.t Srk.Interpretation.interpretation -> Srk.Syntax.symbol list -> t

Best abstraction of a model, restricted to the symbols in the given list.

val formula_of : t -> C.t Srk.Syntax.formula