Module Srk__Interpretation

exception Divide_by_zero

type 'a value = [

| `Real of Srk.QQ.t

| `Bool of bool

| `Fun of ('a, Srk.Syntax.typ_fo) Srk.Syntax.expr

]

type 'a interpretation

An interpretation is a mapping from symbols to values. Interpretations are a collection of concrete bindings (explicit (symbol, value) pairs) combined with an optional abstract binding (a symbol -> value function that can only be inspected by calling).

val pp : Stdlib.Format.formatter -> 'a interpretation -> unit

val empty : 'a Srk.Syntax.context -> 'a interpretation

val wrap : ?⁠symbols:Srk.Syntax.symbol list -> 'a Srk.Syntax.context -> (Srk.Syntax.symbol -> 'a value) -> 'a interpretation

Wrap a function mapping symbols to values in an interpretation. Calls to this function are cached. Symbols belonging to the optional symbol list argument are pre-cached.

val add_real : Srk.Syntax.symbol -> Srk.QQ.t -> 'a interpretation -> 'a interpretation

val add_bool : Srk.Syntax.symbol -> bool -> 'a interpretation -> 'a interpretation

val add_fun : Srk.Syntax.symbol -> ('a, Srk.Syntax.typ_fo) Srk.Syntax.expr -> 'a interpretation -> 'a interpretation

val add : Srk.Syntax.symbol -> 'a value -> 'a interpretation -> 'a interpretation

val real : 'a interpretation -> Srk.Syntax.symbol -> Srk.QQ.t

val bool : 'a interpretation -> Srk.Syntax.symbol -> bool

val value : 'a interpretation -> Srk.Syntax.symbol -> 'a value

val enum : 'a interpretation -> (Srk.Syntax.symbol * [ `Real of Srk.QQ.t | `Bool of bool | `Fun of ('a, Srk.Syntax.typ_fo) Srk.Syntax.expr ]) BatEnum.t

Enumerate the concrete bindings in an interpretation.

val substitute : 'a interpretation -> ('a, 'typ) Srk.Syntax.expr -> ('a, 'typ) Srk.Syntax.expr

Replace constant symbols by their interpretations within an expression. A constant symbol that is not defined within the interpretation is not replaced.

val evaluate_term : 'a interpretation -> ?⁠env:[ `Real of Srk.QQ.t | `Bool of bool ] Srk.Syntax.Env.t -> 'a Srk.Syntax.term -> Srk.QQ.t

val evaluate_formula : 'a interpretation -> ?⁠env:[ `Real of Srk.QQ.t | `Bool of bool ] Srk.Syntax.Env.t -> 'a Srk.Syntax.formula -> bool

val get_context : 'a interpretation -> 'a Srk.Syntax.context

val select_implicant : 'a interpretation -> ?⁠env:[ `Real of Srk.QQ.t | `Bool of bool ] Srk.Syntax.Env.t -> 'a Srk.Syntax.formula -> 'a Srk.Syntax.formula list option

select_implicant srk m ?env phi selects an implicant I of phi such that m,?env |= I |= phi. The implicant I is a list of atomic formulas, which can be destructed using destruct_atom.

val select_ite : 'a interpretation -> ?⁠env:[ `Real of Srk.QQ.t | `Bool of bool ] Srk.Syntax.Env.t -> ('a, 'b) Srk.Syntax.expr -> ('a, 'b) Srk.Syntax.expr * 'a Srk.Syntax.formula list

val destruct_atom : 'a Srk.Syntax.context -> 'a Srk.Syntax.formula -> [ `Comparison of [ `Lt | `Leq | `Eq ] * 'a Srk.Syntax.term * 'a Srk.Syntax.term | `Literal of [ `Pos | `Neg ] * [ `Const of Srk.Syntax.symbol | `Var of int ] ]