Module Srk__Linear

Various operations for the vector space int -> QQ

exception No_solution

Raised for unsolvable systems of linear equations

exception Nonlinear

module ZZVector : sig ... end

Sparse vector with integer entries.

module QQVector : sig ... end

Sparse vector with rational entries.

module QQMatrix : sig ... end

Sparse matrix with rational entries.

val nullspace : QQMatrix.t -> int list -> QQVector.t list

nullspace mat dimensions computes a basis for the vector space { x : mat*x = 0}, projected on to the set of dimensions dimensions. (Note that the nullspace is not finitely generated in int -> QQ, hence the projection).

val solve_exn : QQMatrix.t -> QQVector.t -> QQVector.t

solve_exn mat b computes a rational vector x such that mat*x = b. Raises No_solution if there is no solution.

val solve : QQMatrix.t -> QQVector.t -> QQVector.t option

solve mat b computes a rational vector x such that mat*x = b, if such a vector exists. Otherwise, return None.

val vector_right_mul : QQMatrix.t -> QQVector.t -> QQVector.t

vector_right_mul m v computes m*v

val vector_left_mul : QQVector.t -> QQMatrix.t -> QQVector.t

vector_left_mul v m computes (v^t)*m

val intersect_rowspace : QQMatrix.t -> QQMatrix.t -> QQMatrix.t * QQMatrix.t

Given two matrices A and B, compute matrices C and D such that CA = DB is a basis for the intersection of the rowspaces of A and B.

val pushout : QQMatrix.t -> QQMatrix.t -> QQMatrix.t * QQMatrix.t

Given two matrices A and B, compute matrices C and D such that CA = DB, and for any E and F such that EA = FB, there exists a (unique) U such that UCA = EA (= UDB = FB)

val divide_right : QQMatrix.t -> QQMatrix.t -> QQMatrix.t option

Given two matrices A and B, compute a matrix C such that CB = A (if one exists). C exists when the rowspace of B is contained in the rowspace of A. If A and B are invertible, then C is exactly AB-1.

val divide_left : QQMatrix.t -> QQMatrix.t -> QQMatrix.t option

Given two matrices A and B, compute a matrix C such that BC = A (if one exists). C exists when the columnspace of B is contained in the columnspace of A. If A and B are invertible, then C is exactly B-1A.

val rational_spectral_decomposition : QQMatrix.t -> int list -> (Srk.QQ.t * QQVector.t) list

Compute a list of (rational eigenvalue, generalized eigenvector) pairs of a finite dimensional square submatrix of a given matrix. The submatrix is determined by the list of dimensions given as the second parameter.

val periodic_rational_spectral_decomposition : QQMatrix.t -> int list -> (int * Srk.QQ.t * QQVector.t) list

Compute a list of (period, rational eigenvalue, generalized eigenvector) triples of a finite dimensional square submatrix of a given matrix. The submatrix is determined by the list of dimensions given as the second parameter.

val jordan_chain : QQMatrix.t -> Srk.QQ.t -> QQVector.t -> QQVector.t list

jordan_chain mA lambda v computes the (left) Jordan chain generated by v corresponding to the eigenvalue lambda of mA. This is a list of vectors v0,...,vn such that

v0 = v

for each i < n we have vi(mA-lambda*I) = v+1

vn(mA-lambda*I) = 0

module QQVectorSpace : sig ... end

Affine terms

val sym_of_dim : int -> Srk.Syntax.symbol option

Map a symbol to a dimension. The following equations hold:

sym_of_dim (dim_of_sym sym) = Some sym

sym_of_dim const_dim = None

val dim_of_sym : Srk.Syntax.symbol -> int

Map a dimension to symbol. The following equations hold:

sym_of_dim (dim_of_sym sym) = Some sym

sym_of_dim const_dim = None

val const_dim : int

Dimension for representing the coefficient of the constant 1.

val const_linterm : Srk.QQ.t -> QQVector.t

Representation of a rational number as an affine term. The equation const_of_linterm (const_linterm qq) = Some qq must hold.

val const_of_linterm : QQVector.t -> Srk.QQ.t option

Convert an affine term to a rational number, if possible. The equation const_of_linterm (const_linterm qq) = Some qq must hold.

val linterm_of : 'a Srk.Syntax.context -> 'a Srk.Syntax.term -> QQVector.t

Convert a rational vector representing an affine term. Raises Nonlinear if the input term is non-linear.

val of_linterm : 'a Srk.Syntax.context -> QQVector.t -> 'a Srk.Syntax.term

Convert a rational vector to an affine term. The equation of_linterm srk (linterm_of srk t) = t must hold.

val pp_linterm : 'a Srk.Syntax.context -> Stdlib.Format.formatter -> QQVector.t -> unit

Pretty-print an affine term

val evaluate_linterm : (Srk.Syntax.symbol -> Srk.QQ.t) -> QQVector.t -> Srk.QQ.t

evaluate_linterm env t evaluates the affine term t in the environment env

val linterm_size : QQVector.t -> int

Count the number of dimensions with non-zero coefficients

val term_of_vec : 'a Srk.Syntax.context -> (int -> 'a Srk.Syntax.term) -> QQVector.t -> 'a Srk.Syntax.term

term_of_vec srk t vec creates a term representation of vec by interpreting each dimension i as the term t i.

val evaluate_affine : (int -> Srk.QQ.t) -> QQVector.t -> Srk.QQ.t

Evaluate a vector by interpreting each coordinate according to the given interpretation (the interpretation of const_dim is fixed to be 1).