Module Srk__Linear.QQVectorSpace

Rational vector spaces

type t = QQVector.t list

Vector spaces are represented by a list of basis vectors

val equal : t -> t -> bool

val subspace : t -> t -> bool

subspace a b checks whether a is a subspace of b

val empty : t

Empty vector space

val is_empty : t -> bool

Check whether a given vector space is empty.

val matrix_of : t -> QQMatrix.t

Create a matrix whose rows form a basis for the space

val of_matrix : QQMatrix.t -> t

Given a matrix *with linearly independent rows*, get basis for its rowspace

val mem : t -> QQVector.t -> bool

Given a vector space V and a vector v, does v belong to V

val intersect : t -> t -> t

Intersect two vector spaces

val sum : t -> t -> t

Given vector spaces U and V, compute a basis for the direct sum { u+v : u in U, v in V }

val diff : t -> t -> t

Given vector spaces U and V, compute a basis for a vector space W such that sum (diff U V) W = U. NOTE: this vector space is not unique.

val standard_basis : int -> t

standard_basis n returns standard basis for n-dimensional space

val basis : QQVector.t list -> t

Compute a basis for the vector space spanned by a given set of vectors

val simplify : t -> t

Simplify a basis using Gauss-Jordan elimination

val dimension : t -> int

Find the dimension of the given vector space