Module Srk__Loop.Make

Ordered loop nesting. forests

Parameters

G : G

Signature

type loop

A loop of a graph G is a pair (h,B) consisting of a set of vertices B, and a header vertex h that belongs to B such that subgraph of G induced by B is strongly connected.

type forest = [ `Vertex of G.V.t | `Loop of loop ] list

An ordered loop nesting forest of a graph G consists of a sequence of the strongly connected components of G in topological order, and such that each non-trival SCC C is associated with (1) a header vertex v, and (2) an ordered loop nesting forest for the subgraph of G induced by C \ H. (Note that an SCC consisting of a single vertex with a self-loop is considered a non-trival SCC). An ordered loop nesting forest is simultaneously a loop nesting forest (Ramalingam: On loops, dominators, and dominance frontiers) and a weak topological order (Bourdoncle: Efficient chaotic iteration strategies with widenings).

module VertexSet = BatSet.Make(G.V)

val loop_nest : G.t -> forest

Construct an ordered loop nesting forest, and return the list of roots.

val header : loop -> G.V.t

Header vertex of a loop. If any vertex in the loop has an incoming edge from outside the loop, so does the header.

val body : loop -> VertexSet.t

Set of vertices belonging to a loop.

val children : loop -> forest

children loop produces a list of the vertices and loops for which loop is the least (most deeply nested) that contains it, ordered topologically. Note that children does not contain the loop header.

val cutpoints : forest -> VertexSet.t

Compute a list of all headers of all loops in a loop nesting forest for a graph G. The subgraph obtained by removing the incoming edges to each header vertex is acyclic.

val all_loops : forest -> loop list

Compute a list of all loops in a loop nesting forest. Every pair of loops is either disjoint or one is a proper subloop of the other (a loop (h,B) is a subloop of (h',B') if B is a subset of B' - {h}.

val pp : (Stdlib.Format.formatter -> G.V.t -> unit) -> Stdlib.Format.formatter -> forest -> unit

Pretty-print.