Module Srk__Lts

Linear transition systems. A transition system is linear if its state space is a finite dimensional vector space and its transition relation is a linear space.

module QQMatrix = Srk.Linear.QQMatrix

module QQVector = Srk.Linear.QQVector

module QQVectorSpace = Srk.Linear.QQVectorSpace

type lts = QQMatrix.t * QQMatrix.t

Linear transition systems

val pp : (Stdlib.Format.formatter -> int -> unit) -> Stdlib.Format.formatter -> lts -> unit

val abstract_lts : 'a Srk.Syntax.context -> 'a Srk.TransitionFormula.t -> lts

Find the best LTS abstraction of a transition formula w.r.t. affine simulations

val contains : lts -> lts -> bool

Determine if the transition relation of one LTS is contained within another

val containment_witness : lts -> lts -> Srk.Linear.QQMatrix.t option

Given LTS Ax' = Bx and Cx' = Dx, if Ax' = Bx entails Cx' = Dx, find a matrix M such that MA = C and MB = D; otherwise return None.

module PartialLinearMap : sig ... end

type dlts = PartialLinearMap.t

Deterministic linear transition systems.

val determinize : lts -> dlts * QQMatrix.t

Find the best deterministic abstraction of a linear transition system.

val dlts_inverse_image : QQMatrix.t -> dlts -> lts

Given a dlts T, and a linear map s into the state space of T, compute the inverse image of the transition relation of T under s.

val rational_spectrum_reflection : dlts -> int -> dlts * QQMatrix.t

Find the best abstraction of a DLTS as a DLTS with rational spectrum.

val periodic_rational_spectrum_reflection : dlts -> int -> dlts * QQMatrix.t

Find the best abstraction of a DLTS as a DLTS with periodic rational spectrum.