Module Srk__Polynomial

Polynomials and Grobner bases.

module type Univariate = sig ... end

Signature of univariate polynmials

module MakeUnivariate : functor (R : Srk.Ring.S) -> Univariate with type scalar = R.t

Univariate polynomials over a given ring

module QQX : sig ... end

Univariate polynomials with rational coefficients

module Monomial : sig ... end

Monomials

module type Multivariate = sig ... end

Signature of multivariate polynmials

module MakeMultivariate : functor (R : Srk.Ring.S) -> Multivariate with type scalar = R.t

Multi-variate polynomials over a ring

module QQXs : sig ... end

Multi-variate polynomials with rational coefficients

module Rewrite : sig ... end

Rewrite systems for multi-variate polynomials. A polynomial rewrite system consists of a set of polynomial rewrite rules m1 -> p1, ..., mn -> pn where each mi is a monomial, each pi is a polynomial, and such that mi is greater than any monomial in pi in some admissible order. An admissible order is a total order on monomial such that for all m, n, p, we have: 1. m <= n implies mp <= np 2. m <= mp

module Ideal : sig ... end

A polynomial ideal is a set of polynomials that is closed under addition, and is closed under multiplication by any polynomial.