Module Srk__Polynomial.MakeMultivariate

Multi-variate polynomials over a ring

Parameters

R : Srk.Ring.S

Signature

type t

type scalar = R.t

val equal : t -> t -> bool

val add : t -> t -> t

val negate : t -> t

val mul : t -> t -> t

val zero : t

val one : t

val sub : t -> t -> t

val pp : (Stdlib.Format.formatter -> scalar -> unit) -> (Stdlib.Format.formatter -> int -> unit) -> Stdlib.Format.formatter -> t -> unit

val add_term : scalar -> Monomial.t -> t -> t

add_term k m p is the polynomial k*m + p

val scalar : scalar -> t

Constant polynomial

val of_dim : int -> t

Polynomial consisting of a single variable

val scalar_mul : scalar -> t -> t

val coeff : Monomial.t -> t -> scalar

val pivot : Monomial.t -> t -> scalar * t

val enum : t -> (scalar * Monomial.t) BatEnum.t

val of_enum : (scalar * Monomial.t) BatEnum.t -> t

val of_list : (scalar * Monomial.t) list -> t

val exp : t -> int -> t

Exponentiation by a positive integer

val substitute : (int -> t) -> t -> t

Generalization of polynomial composition -- substitute each dimension for a multivariate polynomial

val mul_monomial : Monomial.t -> t -> t

Multiply a polynomial by a monomial

val div_monomial : t -> Monomial.t -> t option

Divide a polynomial by a monomial

val qr_monomial : t -> Monomial.t -> t * t

Given polynomial p and monomial m, compute polynomials q and r such that p = q*m + r, and such that m does not divide any term in r.

val dimensions : t -> Srk.SrkUtil.Int.Set.t

The set of dimensions that appear in a polynomial

val degree : t -> int

Maximum total degree of a monomial term