Module Srk__Polynomial.MakeUnivariate

Univariate polynomials over a given ring

Parameters

R : Srk.Ring.S

Signature

include Srk.Ring.Vector with type dim = int

type t

type dim = int

type scalar

val equal : t -> t -> bool

Vector equality.

val add : t -> t -> t

Vector addition.

val scalar_mul : scalar -> t -> t

Multiply a vector by a scalar.

val negate : t -> t

Additive inverse of a vector.

val sub : t -> t -> t

Vector subtraction.

val dot : t -> t -> scalar

Inner product: dot u v = sum_i (u_i * v_i).

val zero : t

The vector with all zero entries.

val is_zero : t -> bool

Is the given vector the zero vector?

val add_term : scalar -> dim -> t -> t

add_term a i v adds a to the ith position of v.

val of_term : scalar -> dim -> t

of_term a i is the all-zero vector except has a at position i.

val enum : t -> (scalar * dim) BatEnum.t

Enumerate over all (non-zero) entries of a vector and their position.

val of_enum : (scalar * dim) BatEnum.t -> t

Convert an enumeration into a vector. Satisfies of_enum (enum v) = v, but not necessarily enum (of_enum e) = e; equivalent to add_term a i over all (a,i) in the given enumeration so (1) zero entries are suppressed and (2) if the enumeration contains multiple entries for the samae position, those entries are added.

val of_list : (scalar * dim) list -> t

See of_enum.

val coeff : dim -> t -> scalar

coeff i v finds the entry at the ith position of v.

val pivot : dim -> t -> scalar * t

pivot i v find the entry at the ith position of v, and the result of removing the ith position (replacing it with 0). If (a,v') = pivot i v, then add_term a i v' = v.

val order : t -> int

val mul : t -> t -> t

val one : t

val scalar : scalar -> t

val compose : t -> t -> t

val identity : t

The polynomial p(x) = x

val eval : t -> scalar -> scalar

val exp : t -> int -> t

Exponentiation

val mul_monomial : scalar -> int -> t -> t

mul_monomial k d p multiplies the polynomial p by k * x^d