Module Srk__Polynomial.QQXs

Multi-variate polynomials with rational coefficients

include Multivariate with type scalar = Srk.QQ.t

type t

type scalar = Srk.QQ.t

val equal : t -> t -> bool

val add : t -> t -> t

val negate : t -> t

val mul : t -> t -> t

val zero : t

val one : t

val sub : t -> t -> t

val pp : (Stdlib.Format.formatter -> scalar -> unit) -> (Stdlib.Format.formatter -> int -> unit) -> Stdlib.Format.formatter -> t -> unit

val add_term : scalar -> Monomial.t -> t -> t

add_term k m p is the polynomial k*m + p

val scalar : scalar -> t

Constant polynomial

val of_dim : int -> t

Polynomial consisting of a single variable

val scalar_mul : scalar -> t -> t

val coeff : Monomial.t -> t -> scalar

val pivot : Monomial.t -> t -> scalar * t

val enum : t -> (scalar * Monomial.t) BatEnum.t

val of_enum : (scalar * Monomial.t) BatEnum.t -> t

val of_list : (scalar * Monomial.t) list -> t

val exp : t -> int -> t

Exponentiation by a positive integer

val substitute : (int -> t) -> t -> t

Generalization of polynomial composition -- substitute each dimension for a multivariate polynomial

val mul_monomial : Monomial.t -> t -> t

Multiply a polynomial by a monomial

val div_monomial : t -> Monomial.t -> t option

Divide a polynomial by a monomial

val qr_monomial : t -> Monomial.t -> t * t

Given polynomial p and monomial m, compute polynomials q and r such that p = q*m + r, and such that m does not divide any term in r.

val dimensions : t -> Srk.SrkUtil.Int.Set.t

The set of dimensions that appear in a polynomial

val degree : t -> int

Maximum total degree of a monomial term

val pp : (Stdlib.Format.formatter -> int -> unit) -> Stdlib.Format.formatter -> t -> unit

val compare : t -> t -> int

val of_vec : ?⁠const:int -> Srk.Linear.QQVector.t -> t

Convert a rational vector to a linear polynomial, where each dimension corresponds to a variable except the designated const dimension, which is treated at a constant 1.

val split_linear : ?⁠const:int -> t -> Srk.Linear.QQVector.t * t

Write a polynomial as a sum t + p, where t is a linear term and p is a polynomial in which every monomial has degree >= 2

val vec_of : ?⁠const:int -> t -> Srk.Linear.QQVector.t option

Convert a linear polynomial to a vector, where each dimension corresponds to a variable except the designated const dimension, which is treated at a constant 1. Return None if the polynomial is not linear.

val term_of : 'a Srk.Syntax.context -> (Monomial.dim -> 'a Srk.Syntax.term) -> t -> 'a Srk.Syntax.term

val content : t -> Srk.QQ.t

Greatest common divisor of all coefficients.

val factor_gcd : t -> Srk.QQ.t * Monomial.t * t

Write a polynomial p(x) as c*m(x)*q(x), where c is the content of p, m is the greatest common divisors of all monomials in p, and q is the remainder.

val split_leading : (Monomial.t -> Monomial.t -> [ `Eq | `Lt | `Gt ]) -> t -> Srk.QQ.t * Monomial.t * t

Write a polynomial p(x) as c*m(x) + q(x), where c is the leading coefficient and q is the leading monomial of p (w.r.t. the given monomial order