Module Srk__QQ

Unbounded rationals

type t = Mpqf.t

val pp : Stdlib.Format.formatter -> t -> unit

val show : t -> string

val compare : t -> t -> int

val equal : t -> t -> bool

val leq : t -> t -> bool

val lt : t -> t -> bool

val hash : t -> int

val add : t -> t -> t

val mul : t -> t -> t

val div : t -> t -> t

val idiv : t -> t -> Srk.ZZ.t

val modulo : t -> t -> t

val gcd : t -> t -> t

val lcm : t -> t -> t

val zero : t

val one : t

val negate : t -> t

val inverse : t -> t

val floor : t -> Srk.ZZ.t

val ceiling : t -> Srk.ZZ.t

val sub : t -> t -> t

val exp : t -> int -> t

val numerator : t -> Srk.ZZ.t

val denominator : t -> Srk.ZZ.t

val to_zz : t -> Srk.ZZ.t option

val to_zzfrac : t -> Srk.ZZ.t * Srk.ZZ.t

val to_float : t -> float

val to_int : t -> int option

val of_string : string -> t

val of_int : int -> t

val of_zz : Srk.ZZ.t -> t

val of_frac : int -> int -> t

val of_zzfrac : Srk.ZZ.t -> Srk.ZZ.t -> t

val of_float : float -> t

val min : t -> t -> t

val max : t -> t -> t

val abs : t -> t

Approximation of rationals

val nudge : ?⁠accuracy:int -> t -> t * t

Compute an upper and lower bound for a rational number (with small representations) via truncated continued fractions. The optional accuracy parameter determines the number of terms in the continued fraction (larger => more accurate, smaller => smaller representation)

val nudge_up : ?⁠accuracy:int -> t -> t

Compute an upper bound for a rational number (with a small representation).

val nudge_down : ?⁠accuracy:int -> t -> t

Compute an lower bound for a rational number (with a small representation).