Module Srk__Quantifier

val simsat : 'a Srk.Syntax.context -> 'a Srk.Syntax.formula -> [ `Sat | `Unsat | `Unknown ]

Satisfiability via strategy improvement

val simsat_forward : 'a Srk.Syntax.context -> 'a Srk.Syntax.formula -> [ `Sat | `Unsat | `Unknown ]

Satisfiability via strategy improvement, forwards version

val maximize : 'a Srk.Syntax.context -> 'a Srk.Syntax.formula -> 'a Srk.Syntax.term -> [ `Bounded of Srk.QQ.t | `Infinity | `MinusInfinity | `Unknown ]

Alternating quantifier optimization

val qe_mbp : 'a Srk.Syntax.context -> 'a Srk.Syntax.formula -> 'a Srk.Syntax.formula

Quantifier eliminiation via model-based projection

val mbp : ?⁠dnf:bool -> 'a Srk.Syntax.context -> (Srk.Syntax.symbol -> bool) -> 'a Srk.Syntax.formula -> 'a Srk.Syntax.formula

Model-based projection. If dnf option is set, convert to disjunctive normal form.

val mbp_cover : ?⁠dnf:bool -> 'a Srk.Syntax.context -> (Srk.Syntax.symbol -> bool) -> 'a Srk.Syntax.formula -> 'a Srk.Syntax.formula

Over-approximtate model-based projection. If dnf option is set, convert to disjunctive normal form.

val easy_sat : 'a Srk.Syntax.context -> 'a Srk.Syntax.formula -> [ `Sat | `Unsat | `Unknown ]

Alternating quantifier satisfiability

type 'a strategy

type quantifier_prefix = ([ `Forall | `Exists ] * Srk.Syntax.symbol) list

val pp_strategy : 'a Srk.Syntax.context -> Stdlib.Format.formatter -> 'a strategy -> unit

val show_strategy : 'a Srk.Syntax.context -> 'a strategy -> string

val winning_strategy : 'a Srk.Syntax.context -> quantifier_prefix -> 'a Srk.Syntax.formula -> [ `Sat of 'a strategy | `Unsat of 'a strategy | `Unknown ]

Compute a winning SAT/UNSAT strategy for a formula. The formula is represented in prenex form (quantifier prefix + matrix).

val check_strategy : 'a Srk.Syntax.context -> quantifier_prefix -> 'a Srk.Syntax.formula -> 'a strategy -> bool

Verify that a SAT strategy is in fact a winning strategy. (To verify an UNSAT strategy, just negate the formula & quantifier prefix)

val normalize : 'a Srk.Syntax.context -> 'a Srk.Syntax.formula -> quantifier_prefix * 'a Srk.Syntax.formula

Write a formula in prenex normal form, using constant symbols instead of variables.

val is_presburger_atom : 'a Srk.Syntax.context -> 'a Srk.Syntax.formula -> bool

val local_project_cube : 'a Srk.Syntax.context -> (Srk.Syntax.symbol -> bool) -> 'a Srk.Interpretation.interpretation -> 'a Srk.Syntax.formula list -> 'a Srk.Syntax.formula list

Given an interpretation M, a conjunctive formula cube such that M |= cube, and a predicate p, find a cube cube' expressed over symbols that satisfy p such that M |= cube' |= cube. local_project_cube has a finite image in the sense that for any cube c, the set { local_project_cube srk p m c : m |= c } is finite. local_project_cube assumes a formula in QF_LRA; if not, then cube' may not entail cube, but it is guaranteed to be satisfied by M.