Module Srk__Ring

Operations on rings.

module type S = sig ... end

Signature of rings

module type Vector = sig ... end

Sparse vectors of infinite dimension.

module type Matrix = sig ... end

Sparse matrices of infinite dimension.

module type Map = sig ... end

Minimal signature for maps to support RingMap.

module RingMap : functor (M : Map) -> functor (R : S) -> Vector with type t = R.t M.t and type dim = M.key and type scalar = R.t

Lift a map type over a ring to a left-module

module MakeVector : functor (R : S) -> sig ... end

Sparse vectors of infinite dimension, with entries drawn from a given ring.

module MakeMatrix : functor (R : S) -> sig ... end

Sparse matrices of infinite dimension, with entries drawn from a given ring.

module MakeUltPeriodicSeq : functor (R : S) -> sig ... end

Ultimately periodic sequences. An ultimately periodic sequence is an infinite sequence consisting of a finite transient phase and an infinite periodic phase (with finite period). Ring operations are pointwise: