Module type Srk__Ring.Matrix

Sparse matrices of infinite dimension.

type t

type dim = int

type scalar

type vector

val equal : t -> t -> bool

Matrix equality.

val add : t -> t -> t

Matrix addition.

val scalar_mul : scalar -> t -> t

Multiply a matrix by a scalar.

val mul : t -> t -> t

Matrix multiplication.

val zero : t

Matrix with all zero entries.

val identity : dim list -> t

identity dims is the identity matrix restricted to the dimensions in dims; that is, the matrix with M with Mij = 1 if i in dims and i = j, and Mij = 0 otherwise.

val row : dim -> t -> vector

row i mat finds the ith row vector in mat.

val column : dim -> t -> vector

column j mat finds the jth row vector in mat.

val rowsi : t -> (dim * vector) BatEnum.t

Enumerate over the (non-zero) rows of a matrix and their positions.

val min_row : t -> dim * vector

Find the (non-zero) row at the least position.

val add_row : dim -> vector -> t -> t

add_row i vec mat adds vec (considered as a row vector) to the ith row of mat. Equivalent to adding mat + n, where n is a matrix that is zero everywhere except the ith row, which is vec.

val add_column : dim -> vector -> t -> t

add_column i vec mat adds vec (considered as a column vector) to the jth column of mat. Equivalent to mat + n, where n is a matrix that is zero everywhere except the jth column, which is vec.

val add_entry : dim -> dim -> scalar -> t -> t

add_entry i j a mat adds a to the (i,j)th position of mat.

val pivot : dim -> t -> vector * t

pivot i mat finds the ith row of mat as well as the matrix obtained by removing the ith row (i.e., replacing it with a zero row). If (v,mat') = pivot i mat, then add_row i v mat' = mat.

val pivot_column : dim -> t -> vector * t

pivot_column i mat finds the ith column of mat as well as the matrix obtained by removing the ith column (i.e., replacing it with a zero column). If (v,mat') = pivot i mat, then add_column i v mat' = mat.

val transpose : t -> t

transpose mat is a matrix mat* with mat*ij = matji.

val entry : dim -> dim -> t -> scalar

entry i j mat is the (i,j)th position of mat.

val entries : t -> (dim * dim * scalar) BatEnum.t

Enumerate over all (non-zero) entries of a matrix and their positions

val row_set : t -> Srk.SrkUtil.Int.Set.t

Set of positions corresponding to (non-zero) rows of a matrix

val column_set : t -> Srk.SrkUtil.Int.Set.t

Set of positions corresponding to (non-zero) columns of a matrix

val nb_rows : t -> int

Number of (non-zero) rows

val nb_columns : t -> int

Number of (non-zero) columns

val map_rows : (vector -> vector) -> t -> t

map_rows f mat computes a matrix mat' whose ith row is f mat_i, where mat_i denotes the ith row of mat. Assumes that the function f is zero-strict (f zero = zero).

val vector_right_mul : t -> vector -> vector

vector_right_mul m v computes m*v

val vector_left_mul : vector -> t -> vector

vector_left_mul v m computes (v^t)*m

val of_dense : scalar array array -> t

Compute an equivalent sparse matrix from a dense representation

val dense_of : t -> int -> int -> scalar array array

dense_of mat m n computes the dense m-by-n matrix d where dij = matij.

val of_rows : vector list -> t

Given vectors v0,v1,v2,...,vn create a matrix where row i is vi

val interlace_columns : t -> t -> t

Given two matrices M and N, M and N into a single matrix, using the even columns for M and the odd columns for N