Module Srk__Smt

module Solver : sig ... end

val mk_solver : ?⁠theory:string -> 'a Srk.Syntax.context -> 'a Solver.t

val get_model : ?⁠symbols:Srk.Syntax.symbol list -> 'a Srk.Syntax.context -> 'a Srk.Syntax.formula -> [ `Sat of 'a Srk.Interpretation.interpretation | `Unsat | `Unknown ]

Compute a model of a formula. The model is abstract -- it can be used to evaluate terms, but its bindings may not be enumerated (see Interpretation for more detail).

val get_concrete_model : 'a Srk.Syntax.context -> Srk.Syntax.symbol list -> 'a Srk.Syntax.formula -> [ `Sat of 'a Srk.Interpretation.interpretation | `Unsat | `Unknown ]

Compute a model of a formula, and return an intepretation that binds the specified subset of symbols. If the symbol list contains all symbols of the formula, then the interpretation is a model of the formula.

val is_sat : 'a Srk.Syntax.context -> 'a Srk.Syntax.formula -> [ `Sat | `Unsat | `Unknown ]

val entails : 'a Srk.Syntax.context -> 'a Srk.Syntax.formula -> 'a Srk.Syntax.formula -> [ `Yes | `No | `Unknown ]

val equiv : 'a Srk.Syntax.context -> 'a Srk.Syntax.formula -> 'a Srk.Syntax.formula -> [ `Yes | `No | `Unknown ]

val affine_interpretation : 'a Srk.Syntax.context -> 'a Srk.Interpretation.interpretation -> 'a Srk.Syntax.formula -> [ `Sat of 'a Srk.Interpretation.interpretation | `Unsat | `Unknown ]

Given an a model m and a formula phi such that m |= phi, attempt to compute a new interpretation m' such that m' |= phi, m(x) = m'(x) for all constant symbols and non-real functions, and for all real functions f, m'(f) is affine. Return `Unsat if there is no such m', or `Unknown if the status of the associated SMT query could not be determined.