Module Srk__SolvablePolynomial

Solvable polynomial maps. A polynomial map f : Q^n -> Q^n is solvable if its circular dependencies are linear (e.g. f(x,y) = (x+1,y+x^2) is solvable but f(x,y) = (y^2,x) is not).

module SolvablePolynomialOne : Srk.Iteration.PreDomainWedgeIter

Solvable polynomial maps with eigenvalue 1. Closed forms are polynomial.

module SolvablePolynomial : Srk.Iteration.PreDomainWedgeIter

Solvable polynomial maps without spectral assumptions. Closed forms may use operators from Berg's operational calculus, represented as uninterpreted function symbols that will be allocated upon calling exp.

module SolvablePolynomialPeriodicRational : Srk.Iteration.PreDomainWedgeIter

Solvable polynomial maps with periodic rational eigenvalues. Closed forms are exponential polynomials.

module PresburgerGuard : Srk.Iteration.PreDomain

Extends SolvablePolynomialPeriodicRational with universally quantified precondition expressed over terms with Presurger-definable dynamics.

type 'a dlts_abstraction = {

dlts : Srk.Lts.PartialLinearMap.t;

simulation : 'a Srk.Syntax.term array;

}

Deterministic linear transition systems

module DLTS : Srk.Iteration.PreDomainIter with type 'a t = 'a dlts_abstraction

Deterministic linear transition systems

module DLTSSolvablePolynomial : Srk.Iteration.PreDomainWedge with type 'a t = 'a dlts_abstraction

Solvable polynomial maps with periodic rational eigenvalues, restricted to an algebraic variety, represented as a DLTS with a polynomial simulation.

module DLTSPeriodicRational : sig ... end