Module Srk__SrkSimplify

module RationalTerm : sig ... end

val simplify_terms_rewriter : 'a Srk.Syntax.context -> 'a Srk.Syntax.rewriter

val simplify_terms : 'a Srk.Syntax.context -> 'a Srk.Syntax.formula -> 'a Srk.Syntax.formula

val simplify_term : 'a Srk.Syntax.context -> 'a Srk.Syntax.term -> 'a Srk.Syntax.term

val purify : 'a Srk.Syntax.context -> 'a Srk.Syntax.formula -> 'a Srk.Syntax.formula * ('a, Srk.Syntax.typ_fo) Srk.Syntax.expr Srk.Syntax.Symbol.Map.t

Purify function applications in a ground formula: replace each function application within a formula with a fresh symbol, and return both the resulting formula and a mapping from the fresh symbols to the terms they define.

val partition_implicant : 'a Srk.Syntax.formula list -> 'a Srk.Syntax.formula list list

val simplify_conjunction : 'a Srk.Syntax.context -> 'a Srk.Syntax.formula list -> 'a Srk.Syntax.formula list

Given a list of formulas xs, find a sublist xs' such that the conjunction of all xs is equivalent to conjunction of all xs'. The sublist xs' is not guaranteed to be minimal.

val isolate_linear : 'a Srk.Syntax.context -> Srk.Syntax.symbol -> 'a Srk.Syntax.term -> (Srk.QQ.t * 'a Srk.Syntax.term) option

Given a term t and a constant symbol x, find a coefficient a and a term s such that t is equal to a*x + s (a may be 0). Return None if this is not possible (e.g., the term x*x).

val simplify_dda : 'a Srk.Syntax.context -> 'a Srk.Syntax.formula -> 'a Srk.Syntax.formula

Formula simplification given in Isil Dillig, Thomas Dillig, Alex Aiken: "Small formulas for large programs: on-line constraint simplification in scalable static analysis", SAS 2010.