Module Srk__SrkZ3.Solver

type 'a t

val add : 'a t -> 'a Srk.Syntax.formula list -> unit

val push : 'a t -> unit

val pop : 'a t -> int -> unit

val reset : 'a t -> unit

val check : 'a t -> 'a Srk.Syntax.formula list -> [ `Sat | `Unsat | `Unknown ]

val to_string : 'a t -> string

val get_model : ?⁠symbols:Srk.Syntax.symbol list -> 'a t -> [ `Sat of 'a Srk.Interpretation.interpretation | `Unsat | `Unknown ]

Compute a model of a solver's context. The model is abstract -- it can be used to evaluate terms, but its bindings may not be enumerated (see Interpretation for more detail).

val get_concrete_model : 'a t -> Srk.Syntax.symbol list -> [ `Sat of 'a Srk.Interpretation.interpretation | `Unsat | `Unknown ]

Compute a model of the a solver's context, and return an intepretation that binds the specified subset of symbols. If the symbol list contains all symbols of the formula, then the interpretation is a model of the solver's context.

val get_unsat_core : 'a t -> 'a Srk.Syntax.formula list -> [ `Sat | `Unsat of 'a Srk.Syntax.formula list | `Unknown ]

val get_reason_unknown : 'a t -> string