Module Srk__Syntax

Defines Formulas, Terms, Symbols, Types, and Contexts. Syntactic manipulation of terms and formulas.

type 'a context

A context manages symbols and sharing between expressions. context is a phantom type: the 'a type parameter ensures that expressions don't cross contexts.

module Env : sig ... end

Environments are maps whose domain is a set of free variable symbols. Typically used to travese quantified formulas.

type typ_arith = [

| `TyInt

| `TyReal

]

Types

type typ_fo = [

| `TyInt

| `TyReal

| `TyBool

]

type typ = [

| `TyInt

| `TyReal

| `TyBool

| `TyFun of typ_fo list * typ_fo

]

type typ_bool = [

| `TyBool

]

type 'a typ_fun = [

| `TyFun of typ_fo list * 'a

]

val pp_typ : Stdlib.Format.formatter -> [< typ ] -> unit

Symbols

type symbol

Symbols are represented as non-negative integers, but the type definition is hidden. The isomorphism is witnessed by int_of_symbol and symbol_of_int.

val mk_symbol : 'a context -> ?⁠name:string -> typ -> symbol

Create a fresh symbol. Multiple calls to mk_symbol with the same context, name, and type will always result in different symbols.

val register_named_symbol : 'a context -> string -> typ -> unit

Register a named symbol. The strings identifying named symbols must be unique. The symbol associated with a name can be retrieved with get_named_symbol.

val is_registered_name : 'a context -> string -> bool

Test if a name is already associated with a symbol

val get_named_symbol : 'a context -> string -> symbol

Retrieve the symbol associated with a given name. Raises Not_found if there is no such symbol.

val symbol_name : 'a context -> symbol -> string option

Retrieve the name of a named symbol. Evaluates to None for ordinary (non-named) symbols.

val pp_symbol : 'a context -> Stdlib.Format.formatter -> symbol -> unit

val typ_symbol : 'a context -> symbol -> typ

val show_symbol : 'a context -> symbol -> string

val int_of_symbol : symbol -> int

val symbol_of_int : int -> symbol

val compare_symbol : symbol -> symbol -> int

module Symbol : sig ... end

Expressions

type ('a, +'typ) expr

type 'a term = ('a, typ_arith) expr

type 'a formula = ('a, typ_bool) expr

val compare_expr : ('a, 'typ) expr -> ('a, 'typ) expr -> int

val compare_formula : 'a formula -> 'a formula -> int

val compare_term : 'a term -> 'a term -> int

val destruct : 'a context -> ('a, 'b) expr -> [ `Real of Srk.QQ.t | `App of symbol * ('a, typ_fo) expr list | `Var of int * typ_arith | `Add of 'a term list | `Mul of 'a term list | `Binop of [ `Div | `Mod ] * 'a term * 'a term | `Unop of [ `Floor | `Neg ] * 'a term | `Ite of 'a formula * ('a, 'b) expr * ('a, 'b) expr | `Tru | `Fls | `And of 'a formula list | `Or of 'a formula list | `Not of 'a formula | `Quantify of [ `Exists | `Forall ] * string * typ_fo * 'a formula | `Atom of [ `Eq | `Leq | `Lt ] * 'a term * 'a term | `Proposition of [ `Var of int | `App of symbol * ('b, typ_fo) expr list ] ]

val expr_typ : 'a context -> ('a, 'b) expr -> typ

val free_vars : ('a, 'b) expr -> (int, typ_fo) BatHashtbl.t

val size : ('a, 'b) expr -> int

val mk_const : 'a context -> symbol -> ('a, 'typ) expr

val mk_app : 'a context -> symbol -> ('a, 'b) expr list -> ('a, 'typ) expr

val mk_var : 'a context -> int -> typ_fo -> ('a, 'typ) expr

val mk_ite : 'a context -> 'a formula -> ('a, 'typ) expr -> ('a, 'typ) expr -> ('a, 'typ) expr

Create an if-then-else expression.

val mk_if : 'a context -> 'a formula -> 'a formula -> 'a formula

Create an implication formula.

val mk_iff : 'a context -> 'a formula -> 'a formula -> 'a formula

Create an if-and-only-if formula

val substitute : 'a context -> (int -> ('a, 'b) expr) -> ('a, 'typ) expr -> ('a, 'typ) expr

substitute srk subst exp replaces each occurrence of a variable symbol with De Bruijn i with the expression subst i. If subst i contains free variables, capture is avoided.

val substitute_const : 'a context -> (symbol -> ('a, 'b) expr) -> ('a, 'typ) expr -> ('a, 'typ) expr

substitute_const srk subst exp replaces each occurrence of a constant symbol s with the expression subst s. If subst s contains free variables, capture is avoided. Function symbols are not affected.

val substitute_map : 'a context -> ('a, 'b) expr Symbol.Map.t -> ('a, 'typ) expr -> ('a, 'typ) expr

substitute_map srk subst exp replaces each occurrence of a constant symbol s in the domain of the map subst with the expression subst s. If subst s contains free variables, capture is avoided. Function symbols are not affected.

val substitute_sym : 'a context -> (symbol -> ('a, 'b) expr) -> ('a, 'typ) expr -> ('a, 'typ) expr

substitute_sym srk subst exp replaces each occurrence of a an application f(e_0,...,e_n) with the result of replacing the De Bruijn indices 0, ..., n with e_0, ..., e_n in the expression subst f. If subst f contains free variables (beyond 0,...,n), capture is avoided. Constant symbols are treated as nullary function applications, and so are also replaced according to subst.

val fold_constants : (symbol -> 'a -> 'a) -> ('b, 'c) expr -> 'a -> 'a

val symbols : ('a, 'b) expr -> Symbol.Set.t

Expression rewriting

type 'a rewriter = ('a, typ_fo) expr -> ('a, typ_fo) expr

A rewriter is a function that transforms an expression into another. The transformation should preserve types; if not, rewrite will fail.

val rewrite : 'a context -> ?⁠down:'a rewriter -> ?⁠up:'a rewriter -> ('a, 'typ) expr -> ('a, 'typ) expr

Rewrite an expression. The down rewriter is applied to each expression going down the expression tree, and the up rewriter is applied to each expression going up the tree.

val nnf_rewriter : 'a context -> 'a rewriter

Convert to negation normal form (down pass).

module Expr : sig ... end

Terms

type ('a, 'b) open_term = [

| `Real of Srk.QQ.t

| `App of symbol * ('b, typ_fo) expr list

| `Var of int * typ_arith

| `Add of 'a list

| `Mul of 'a list

| `Binop of [ `Div | `Mod ] * 'a * 'a

| `Unop of [ `Floor | `Neg ] * 'a

| `Ite of 'b formula * 'a * 'a

]

val mk_add : 'a context -> 'a term list -> 'a term

val mk_mul : 'a context -> 'a term list -> 'a term

val mk_div : 'a context -> 'a term -> 'a term -> 'a term

val mk_pow : 'a context -> 'a term -> int -> 'a term

val mk_idiv : 'a context -> 'a term -> 'a term -> 'a term

C99 integer division. Equivalent to truncate(x/y).

val mk_mod : 'a context -> 'a term -> 'a term -> 'a term

val mk_real : 'a context -> Srk.QQ.t -> 'a term

val mk_zz : 'a context -> Srk.ZZ.t -> 'a term

val mk_int : 'a context -> int -> 'a term

val mk_zero : 'a context -> 'a term

val mk_one : 'a context -> 'a term

val mk_floor : 'a context -> 'a term -> 'a term

val mk_ceiling : 'a context -> 'a term -> 'a term

val mk_truncate : 'a context -> 'a term -> 'a term

truncate(t) removes the fractional part of t (rounding it towards 0).

val mk_neg : 'a context -> 'a term -> 'a term

Unary negation

val mk_sub : 'a context -> 'a term -> 'a term -> 'a term

Subtraction

val term_typ : 'a context -> 'a term -> typ_arith

module Term : sig ... end

Formulas

type ('a, 'b) open_formula = [

| `Tru

| `Fls

| `And of 'a list

| `Or of 'a list

| `Not of 'a

| `Quantify of [ `Exists | `Forall ] * string * typ_fo * 'a

| `Atom of [ `Eq | `Leq | `Lt ] * 'b term * 'b term

| `Proposition of [ `Var of int | `App of symbol * ('b, typ_fo) expr list ]

| `Ite of 'a * 'a * 'a

]

val mk_forall : 'a context -> ?⁠name:string -> typ_fo -> 'a formula -> 'a formula

val mk_exists : 'a context -> ?⁠name:string -> typ_fo -> 'a formula -> 'a formula

val mk_forall_const : 'a context -> symbol -> 'a formula -> 'a formula

Replace a constant symbol by a universally quantified variable.

val mk_exists_const : 'a context -> symbol -> 'a formula -> 'a formula

Replace a constant symbol by an existentially quantified variable.

val mk_forall_consts : 'a context -> (symbol -> bool) -> 'a formula -> 'a formula

Replace all constant symbols that do not satisfy the given predicate by universally quantified variables.

val mk_exists_consts : 'a context -> (symbol -> bool) -> 'a formula -> 'a formula

Replace all constant symbols that do not satisfy the given predicate by existentially quantified variables.

val mk_and : 'a context -> 'a formula list -> 'a formula

val mk_or : 'a context -> 'a formula list -> 'a formula

val mk_not : 'a context -> 'a formula -> 'a formula

val mk_eq : 'a context -> 'a term -> 'a term -> 'a formula

val mk_lt : 'a context -> 'a term -> 'a term -> 'a formula

val mk_leq : 'a context -> 'a term -> 'a term -> 'a formula

val mk_true : 'a context -> 'a formula

val mk_false : 'a context -> 'a formula

val eliminate_ite : 'a context -> 'a formula -> 'a formula

Given a formula phi, compute an equivalent formula without if-then-else terms. eliminate-ite does not introduce new symbols or quantifiers.

val pp_smtlib2 : ?⁠env:string Env.t -> 'a context -> Stdlib.Format.formatter -> 'a formula -> unit

Print a formula as a satisfiability query in SMTLIB2 format. The query includes function declarations and (check-sat).

val pp_expr_unnumbered : ?⁠env:string Env.t -> 'a context -> Stdlib.Format.formatter -> ('a, 'b) expr -> unit

Print an expression. This variant of pp_expr avoids printing a symbol number (e.g., "x:5") for a symbol S (i.e., a program variable or function name) if there does not exist any other symbol in the expression that has the same name as S.

module Formula : sig ... end

Contexts

module type Context = sig ... end

module MakeContext : functor () Context

module MakeSimplifyingContext : functor () Context

Create a context which simplifies expressions on the fly

module Infix : functor (C : sig ... end) -> sig ... end

module ContextTable : sig ... end

A context table is a hash table mapping contents to values. If a context is garbage collected, the corresponding entry in the table will be removed. The values stored in a context table should have pointers back to their associated context.