Module Srk__TransitionSystem.Make

Parameters

C : sig ... end

Var : sig ... end

T : sig ... end

Signature

type vertex = int

type transition = T.t

type t = transition label Srk.WeightedGraph.t

type query = T.t Srk.WeightedGraph.RecGraph.weight_query

module VarSet : BatSet.S with type VarSet.elt = Var.t

val empty : t

Create an empty transition system.

val mk_query : ?⁠delay:int -> t -> vertex -> (module Srk.WeightedGraph.AbstractWeight with type weight = transition) -> query

Create a query structure. The optional delay parameter specifies the widening delay to use during summary computation.

val path_weight : query -> vertex -> transition

Over-approximate the sum of the weights of all paths from the query's source vertex to the given target vertex.

val call_weight : query -> (vertex * vertex) -> transition

Over-approximate the sum of the weights of all paths between two given vertices.

val remove_temporaries : t -> t

Project out local variables from each transition that are referenced only by that transition.

val forward_invariants_ivl : t -> vertex -> (vertex * C.t Srk.Syntax.formula) list

Compute interval invariants for each loop header of a transition system. The invariant computed for a loop is defined only over the variables read or written to by the loop.

val forward_invariants_ivl_pa : C.t Srk.Syntax.formula list -> t -> vertex -> (vertex * C.t Srk.Syntax.formula) list

Compute interval-and-predicate invariants for each loop header of a transition system. The invariant computed for a loop is defined only over the variables read or written to by the loop.

val simplify : (vertex -> bool) -> t -> t

Simplify a transition system by contracting vertices that do not satisfy the given predicate. Simplification does not guarantee that all such vertices are contracted. In particular, simplification will not contract vertices with loops or vertices adjacent to call edges.

val loop_headers_live : t -> (int * VarSet.t) list

Given a transition system and entry, compute a set of loop headers along with the set of variables that are read within the body of the associated loop

module type AbstractDomain = Srk.Abstract.MakeAbstractRSY(C).Domain

module type IncrAbstractDomain = sig ... end

module LiftIncr : functor (A : AbstractDomain) -> IncrAbstractDomain with type t = A.t

module ProductIncr : functor (A : IncrAbstractDomain) -> functor (B : IncrAbstractDomain) -> IncrAbstractDomain with type t = A.t * B.t

val forward_invariants : (module IncrAbstractDomain with type t = 'a) -> t -> vertex -> vertex -> 'a

forward_invariants d ts entry computes invariants for ts starting from the node entry using the abstract domain d