Module Srk__WeightedGraph.RecGraph

A recursive graph is a graph with two types of edges: simple edges and call edges. Each call edge designates an entry vertex and exit vertex, and can be interpreted a set of paths that begin and entry and end at exit.

type t

type call = vertex * vertex

type query

A query is an intermediate structure for perfoming path expression queries. After creating a recursive graph, a query can be constructed using mk_query and accessed using pathexpr.

type 'a weight_query

A weight query is an intermediate structure for perfoming path weight queries.

exception No_summary of call

module CallGraph : sig ... end

The callgraph of a recursive graph has calls as vertices, and an edge c1 to c2 iff a call to c2 appears on a path in c1. A callgraph can be constructed using mk_callgraph.

val empty : unit -> t

val add_vertex : t -> vertex -> t

val add_edge : t -> vertex -> vertex -> t

val add_call_edge : t -> vertex -> call -> vertex -> t

val mk_query : t -> vertex -> query

Create a query for computing path expressions beginning at the designated source vertex.

val pathexpr : query -> vertex -> Srk.Pathexpr.nested Srk.Pathexpr.t

Find a nested path expression recognizing every interprocedural path from the query's source vertex to the given target vertex. Every complete entry-to-call edge path is enclosed in a segment.

val call_pathexpr : query -> call -> Srk.Pathexpr.simple Srk.Pathexpr.t

Find a simple path expression recognizing every intraprocedural path through a given call. The call must appear on at least one of the call edges of the recursive graph of the query.

val omega_pathexpr : query -> Srk.Pathexpr.nested Srk.Pathexpr.omega

Find an omega regular expression recognizing every infinite interprocedural path beginning at the query's source vertex.

val mk_callgraph : query -> CallGraph.t

val mk_weight_query : query -> 'a Srk.Pathexpr.nested_algebra -> 'a weight_query

Create a weighted query for computing path weights. The supplied algebra is used to assign weights to simple edges; the query maintains summaries for each call that are used to assign weights to call edges.

val summarize_iterative : query -> 'a Srk.Pathexpr.nested_algebra -> ?⁠delay:int -> (module AbstractWeight with type weight = 'a) -> 'a weight_query

Build call summaries via successive approximation.

Find the sum of weights of all interprocedural paths beginning at the query's source vertex and ending at a given target.

val path_weight : 'a weight_query -> vertex -> 'a

Find the sum of weights of all interprocedural paths beginning at the query's source vertex and ending at a given target.

val call_weight : 'a weight_query -> call -> 'a

Find the sum of weights of all intraprocedural paths through a given call.

val get_summary : 'a weight_query -> call -> 'a

val set_summary : 'a weight_query -> call -> 'a -> unit

val omega_path_weight : 'a weight_query -> ('a, 'b) Srk.Pathexpr.omega_algebra -> 'b

Find the sum of weights of all infinite interprocedural paths beginning at the query's source vertex.